+++ ui-core/src/index.ts
@@ -18,6 +18,7 @@ export * from './types/address'
 export * from './types/kits'
 export * from './types/user'
 export * from './types/family'
+export * from './types/logging'
 
 export * from './autoSaveUtils'
 export * from './reactUtils'
